elsai-poc/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
sqlalchemy==2.0.36
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
anthropic==0.39.0
pillow==11.0.0
pytesseract==0.3.13
python-dotenv==1.0.1
